requirements.txt
Django==6.*
psycopg2-binary==2.9.*
whitenoise==6.8.*
Pillow==12.1.*
django-json-widget
openai>=1.109.1
google-genai>=1.65.0
